fix: eBPF attribute should be false when there are no programs
Currently, eBPF attribute is always set to true since `link.QueryPrograms`
mostly returns a successful call. This leads to the program thinking that is has
"some" BPF programs since the code does not check for an empty value.

This eventually trickles down to failures in "validate bpf filter attributes".
The attribute is always true so we never wait and continue checking for the
existence of `tcFilterNames` which may not yet have been updated by the driver.