fix: Check for existence of interface attributes before referring to them in tests
This is now required because not all devices have the exact same attributes.
E.g. PCI Devices whose network interface is moved to the pod's network namespace
don't have the network interface related attributes like:
    - dra.net/alias:
    - dra.net/ebpf:
    - dra.net/encapsulation:
    - dra.net/ifName:
    - dra.net/ipv4:
    - dra.net/mac:
    - dra.net/mtu:
    - gce.dra.net/networkName:
    - gce.dra.net/networkProjectNumber:
    - gce.dra.net/rdmaType

This means that the user needs to check for their existence if referring to them.